requirements.txt
oci>=2.90.0
python-dotenv>=0.19.0